+++ dali/internal/event/actors/custom-actor-internal.h
@@ -70,7 +70,7 @@ private:
   /**
    * @copydoc Internal::Actor::OnStageConnectionExternal
    */
-  virtual void OnStageConnectionExternal( unsigned int depth )
+  virtual void OnStageConnectionExternal( unsigned short depth )
   {
     mImpl->OnStageConnection( depth );
   }
